+++ ioi_utils.py
@@ -46,6 +46,9 @@ for head in CIRCUIT["previous token"]:
     RELEVANT_TOKENS[head] = ["S+1", "and"]
 
 def get_heads_circuit(ioi_dataset, excluded_classes=["calibration"], mlp0=False):
+    for excluded_class in excluded_classes:
+        assert excluded_class in CIRCUIT.keys()
+
     heads_to_keep = {}
 
     for circuit_class in CIRCUIT.keys():
@@ -62,6 +65,63 @@ def get_heads_circuit(ioi_dataset, excluded_classes=["calibration"], mlp0=False)
         return heads_to_keep, mlps_to_keep
 
     return heads_to_keep
+
+
+def do_circuit_extraction(
+    heads_to_remove=None,  # {(2,3) : List[List[int]]: dimensions dataset_size * datapoint_length
+    mlps_to_remove=None,  # {2: List[List[int]]: dimensions dataset_size * datapoint_length
+    heads_to_keep=None,  # as above for heads
+    mlps_to_keep=None,  # as above for mlps
+    ioi_dataset=None,
+    model=None,
+):
+    """
+    if `ablate` then ablate all `heads` and `mlps`
+        and keep everything else same
+    otherwise, ablate everything else
+        and keep `heads` and `mlps` the same
+    """
+
+    # check if we are either in keep XOR remove move from the args
+    ablation, heads, mlps = get_circuit_replacement_hook(
+        heads_to_remove=heads_to_remove,  # {(2,3) : List[List[int]]: dimensions dataset_size * datapoint_length
+        mlps_to_remove=mlps_to_remove,  # {2: List[List[int]]: dimensions dataset_size * datapoint_length
+        heads_to_keep=heads_to_keep,  # as above for heads
+        mlps_to_keep=mlps_to_keep,  # as above for mlps
+        ioi_dataset=ioi_dataset,
+        model=model,
+    )
+
+    metric = ExperimentMetric(
+        metric=logit_diff_target, dataset=ioi_dataset.text_prompts, relative_metric=False
+    )  # TODO make dummy metric
+
+    config = AblationConfig(
+        abl_type="custom",
+        abl_fn=ablation,
+        mean_dataset=ioi_dataset.text_prompts,  # TODO nb of prompts useless ?
+        target_module="attn_head",
+        head_circuit="result",
+        cache_means=True,  # circuit extraction *has* to cache means. the get_mean reset the
+        verbose=True,
+    )
+    abl = EasyAblation(
+        model,
+        config,
+        metric,
+        semantic_indices=ioi_dataset.sem_tok_idx,
+        mean_by_groups=True,  # TO CHECK CIRCUIT BY GROUPS
+        groups=ioi_dataset.groups,
+        blue_pen=False,
+    )
+    model.reset_hooks()
+
+    for layer, head in heads.keys():
+        model.add_hook(*abl.get_hook(layer, head))
+    for layer in mlps.keys():
+        model.add_hook(*abl.get_hook(layer, head=None, target_module="mlp"))
+
+    return model, abl
